+++ packages/cspell-dictionary/src/SpellingDictionary/Typos/index.ts
@@ -1,0 +1,3 @@
+export type { TypoEntry, TyposDef } from './typos';
+export { parseTyposFile, parseTyposLine, processEntriesToTyposDef } from './typosParser';
+export { createTyposDef, extractAllSuggestions } from './util';