+++ packages/cspell-pipe/src/sync/index.ts
@@ -9,6 +9,7 @@ export {
     opJoinStringsSync as opJoinStrings,
     opLastSync as opLast,
     opMapSync as opMap,
+    opReduceSync as opReduce,
     opSkipSync as opSkip,
     opTakeSync as opTake,
     opTapSync as opTap,
@@ -16,3 +17,4 @@ export {
 } from '../operators';
 export type { OperatorSync as Operator } from '../operators';
 export { pipeSync as pipe, pipeSync } from '../pipe';
+export { reduceSync as reduce } from '../reduce';