+++ packages/cspell-trie-lib/src/lib/SimpleDictionaryParser.ts
@@ -64,7 +64,7 @@ export interface ParseDictionaryOptions {
     splitSeparator: RegExp | string;
 }
 
-const RegExpSplit = /(?<!\\)[\s,;]/g;
+const RegExpSplit = /[\s,;]/g;
 
 const _defaultOptions: ParseDictionaryOptions = {
     commentCharacter: LINE_COMMENT,
@@ -215,8 +215,9 @@ export function createDictionaryLineParserMapper(options?: Partial<ParseDictiona
                     line.indexOf('"') >= 0
                         ? line.replace(/".*?"/g, (quoted) => ' ' + quoted.replace(/(\s)/g, '\\$1') + ' ')
                         : line;
-                const words = lineEscaped.split(splitSeparator);
-                yield* words.map((escaped) => escaped.replace(/\\(\s)/g, '$1'));
+
+                const words = splitLine(lineEscaped, splitSeparator);
+                yield* words.map((escaped) => escaped.replace(/\\/g, ''));
                 if (!splitKeepBoth) continue;
             }
             yield line;
@@ -288,3 +289,24 @@ function mergeOptions(
     }
     return opt;
 }
+
+const RegExpToEncode = /\\([\s,;])/g;
+const RegExpDecode = /<<(%[\da-f]{2})>>/gi;
+
+function encodeLine(line: string): string {
+    return line.replace(RegExpToEncode, (_, v) => '<<' + encodeURIComponent(v) + '>>');
+}
+
+function decodeLine(line: string): string {
+    return line.replace(RegExpDecode, (_, v) => '\\' + decodeURIComponent(v));
+}
+
+function splitLine(line: string, regExp: RegExp | string): string[] {
+    return encodeLine(line)
+        .split(regExp)
+        .map((line) => decodeLine(line));
+}
+
+export const __testing__ = {
+    splitLine,
+};