+++ rfc/rfc-0001 suggestions/src/config-definitions.ts
@@ -24,7 +24,9 @@ const exampleDef: DictionaryDefinitionSuggestions = {
 /*********************/
 
 type FlagWordNoSuggestions = string;
-type FlagWordWithSuggestions = [forbidWord: string, suggestion: string, ...otherSuggestions: string[]];
+type FlagWordWithSuggestionsArray = [forbidWord: string, suggestion: string, ...otherSuggestions: string[]];
+type FlagWordWithSuggestionsObj = Record<string, string | string[]>;
+type FlagWordWithSuggestions = FlagWordWithSuggestionsArray | FlagWordWithSuggestionsObj;
 type FlagWord = FlagWordNoSuggestions | FlagWordWithSuggestions;
 export type FlagWords = FlagWord[];
 
@@ -33,14 +35,15 @@ export interface NewBaseSettings extends Omit<BaseSetting, 'flagWords'> {
     flagWords?: FlagWords;
 }
 
+// cspell:ignore akcent alusion
 const exampleFlagWords: NewBaseSettings = {
-    flagWords: ['crap', ['hte', 'the']],
+    flagWords: ['crap', ['hte', 'the'], { akcent: 'accent' }, { alusion: ['allusion', 'illusion'] }],
 };
 
 type TermForbid = false;
 type TermIgnore = null;
 type TermWord = true;
-type TermTypo = string[];
+type TermTypo = string[] | string;
 
 const TermForbid: TermForbid = false;
 const TermWord: TermWord = true;