+++ packages/cspell/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 6.14.3 (2022-11-17)
+
+### Bug Fixes
+
+* trie-lib - fix issue with reference radix. ([#3849](https://github.com/streetsidesoftware/cspell/issues/3849)) ([73af697](https://github.com/streetsidesoftware/cspell/commit/73af6978c53f2fd1f4cd528fd84e9d52a4149d5b))
+
+## 6.14.2 (2022-11-11)
+
+**Note:** Version bump only for package cspell
+
 ## 6.14.1 (2022-11-07)
 
 **Note:** Version bump only for package cspell